requirements.txt
fastapi
uvicorn
pandas
requests
python-dotenv
openai
flask
